+++ processors/src/main/java/jp/wasabeef/fresco/processors/gpu/ContrastFilterPostprocessor.java
@@ -1,14 +1,14 @@
 package jp.wasabeef.fresco.processors.gpu;
 
 /**
- * Copyright (C) 2017 Wasabeef
- *
+ * Copyright (C) 2020 Wasabeef
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * <p>
  * http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,16 +17,18 @@ package jp.wasabeef.fresco.processors.gpu;
  */
 
 import android.content.Context;
+
 import com.facebook.cache.common.CacheKey;
 import com.facebook.cache.common.SimpleCacheKey;
-import jp.co.cyberagent.android.gpuimage.GPUImageContrastFilter;
+
+import jp.co.cyberagent.android.gpuimage.filter.GPUImageContrastFilter;
 
 /**
  * contrast value ranges from 0.0 to 4.0, with 1.0 as the normal level
  */
 public class ContrastFilterPostprocessor extends GPUFilterPostprocessor {
 
-  private float contrast;
+  private final float contrast;
 
   public ContrastFilterPostprocessor(Context context) {
     this(context, 1.0f);
@@ -40,7 +42,8 @@ public class ContrastFilterPostprocessor extends GPUFilterPostprocessor {
     filter.setContrast(contrast);
   }
 
-  @Override public CacheKey getPostprocessorCacheKey() {
+  @Override
+  public CacheKey getPostprocessorCacheKey() {
     return new SimpleCacheKey("contrast=" + contrast);
   }
 }