+++ processors/src/main/java/jp/wasabeef/fresco/processors/gpu/SwirlFilterPostprocessor.java
@@ -1,14 +1,14 @@
 package jp.wasabeef.fresco.processors.gpu;
 
 /**
- * Copyright (C) 2017 Wasabeef
- *
+ * Copyright (C) 2020 Wasabeef
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * <p>
  * http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -18,18 +18,20 @@ package jp.wasabeef.fresco.processors.gpu;
 
 import android.content.Context;
 import android.graphics.PointF;
+
 import com.facebook.cache.common.CacheKey;
 import com.facebook.cache.common.SimpleCacheKey;
-import jp.co.cyberagent.android.gpuimage.GPUImageSwirlFilter;
+
+import jp.co.cyberagent.android.gpuimage.filter.GPUImageSwirlFilter;
 
 /**
  * Creates a swirl distortion on the image.
  */
 public class SwirlFilterPostprocessor extends GPUFilterPostprocessor {
 
-  private float radius;
-  private float angle;
-  private PointF center;
+  private final float radius;
+  private final float angle;
+  private final PointF center;
 
   public SwirlFilterPostprocessor(Context context) {
     this(context, 0.5f, 1.0f, new PointF(0.5f, 0.5f));
@@ -37,7 +39,7 @@ public class SwirlFilterPostprocessor extends GPUFilterPostprocessor {
 
   /**
    * @param radius from 0.0 to 1.0, default 0.5
-   * @param angle minimum 0.0, default 1.0
+   * @param angle  minimum 0.0, default 1.0
    * @param center default (0.5, 0.5)
    */
   public SwirlFilterPostprocessor(Context context, float radius, float angle, PointF center) {
@@ -52,8 +54,9 @@ public class SwirlFilterPostprocessor extends GPUFilterPostprocessor {
     filter.setCenter(this.center);
   }
 
-  @Override public CacheKey getPostprocessorCacheKey() {
+  @Override
+  public CacheKey getPostprocessorCacheKey() {
     return new SimpleCacheKey(
-        "radius=" + radius + ",angle=" + angle + ",center=" + center.toString());
+      "radius=" + radius + ",angle=" + angle + ",center=" + center.toString());
   }
 }