+++ processors/src/main/java/jp/wasabeef/fresco/processors/gpu/ToonFilterPostprocessor.java
@@ -1,14 +1,14 @@
 package jp.wasabeef.fresco.processors.gpu;
 
 /**
- * Copyright (C) 2017 Wasabeef
- *
+ * Copyright (C) 2020 Wasabeef
+ * <p>
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- *
+ * <p>
  * http://www.apache.org/licenses/LICENSE-2.0
- *
+ * <p>
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -17,9 +17,11 @@ package jp.wasabeef.fresco.processors.gpu;
  */
 
 import android.content.Context;
+
 import com.facebook.cache.common.CacheKey;
 import com.facebook.cache.common.SimpleCacheKey;
-import jp.co.cyberagent.android.gpuimage.GPUImageToonFilter;
+
+import jp.co.cyberagent.android.gpuimage.filter.GPUImageToonFilter;
 
 /**
  * The threshold at which to apply the edges, default of 0.2.
@@ -28,8 +30,8 @@ import jp.co.cyberagent.android.gpuimage.GPUImageToonFilter;
  */
 public class ToonFilterPostprocessor extends GPUFilterPostprocessor {
 
-  private float threshold;
-  private float quantizationLevels;
+  private final float threshold;
+  private final float quantizationLevels;
 
   public ToonFilterPostprocessor(Context context) {
     this(context, .2f, 10.0f);
@@ -46,8 +48,9 @@ public class ToonFilterPostprocessor extends GPUFilterPostprocessor {
     filter.setQuantizationLevels(this.quantizationLevels);
   }
 
-  @Override public CacheKey getPostprocessorCacheKey() {
+  @Override
+  public CacheKey getPostprocessorCacheKey() {
     return new SimpleCacheKey(
-        "threshold=" + threshold + ",quantizationLevels=" + quantizationLevels);
+      "threshold=" + threshold + ",quantizationLevels=" + quantizationLevels);
   }
 }